requirements.txt
streamlit
torch
torchvision
pillow
numpy
opencv-python
plotly
pdf2image
pymupdf
matplotlib
seaborn
scikit-learn
datasets
